requirements.txt
# Social Media API clients
requests>=2.31.0
tweepy>=4.14.0

# Google Sheets
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
python-dateutil>=2.8.2
pytz>=2024.1

# CLI & Scheduling
click>=8.1.7
schedule>=1.2.1

